Deixei a tela home mais responsiva
