+++ projeto-pokedex/src/components/Home/HomeStyle.js
@@ -4,6 +4,7 @@ export const ContainerGeral = styled.div`
   width: 100%;
   min-height: 120vh;
   display: flex;
+  justify-content: center;
   flex-wrap: wrap;
   background-color: #5e5e5e;
   padding: 60px 25px;
@@ -28,30 +29,7 @@ export const Title = styled.h1`
   padding-bottom: 35px;
 `;
 
-export const ButtonsDiv = styled.div`
-  display: flex;
-  width: 100%;
-  height: 100%;
-  justify-content: center;
-  padding: 12px 0px;
-  flex-wrap: wrap;
-  @media only screen and (max-width: 480px) {
-    /* flex-wrap: wrap; */
-  }
-`;
-
-export const Button = styled.button`
-font-family: "Poppins";
-  color: white;
-  background: none;
-  border: 1px solid black;
-  min-width: 32px;
-  min-height: 32px;
-  padding: 0px 4px;
-  transition: 0.5s ease;
-  cursor: pointer;
-  :hover {
-    background-color: #fff;
-    color: #000;
-  }
-`;
+export const Loading = styled.h1`
+color: #fff;
+font-size: 6rem;
+`
